refactor: Refactoring test code with changed API
