+++ src/test/kotlin/app/hyuabot/backend/building/RoomServiceTest.kt
@@ -1,6 +1,5 @@
-package app.hyuabot.backend.room
+package app.hyuabot.backend.building
 
-import app.hyuabot.backend.building.RoomService
 import app.hyuabot.backend.building.domain.RoomRequest
 import app.hyuabot.backend.building.exception.BuildingNotFoundException
 import app.hyuabot.backend.building.exception.DuplicateRoomException
@@ -18,7 +17,6 @@ import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import java.util.Optional
 import kotlin.test.assertEquals
 
 @ExtendWith(MockitoExtension::class)
@@ -31,23 +29,6 @@ class RoomServiceTest {
 
     @InjectMocks
     lateinit var roomService: RoomService
-
-    @Test
-    @DisplayName("강의실 목록 조회 테스트")
-    fun getRoomListTest() {
-        whenever(roomRepository.findAll()).thenReturn(
-            listOf(
-                Room(seq = 1, buildingName = "Building A", number = "101", name = "Room 101"),
-                Room(seq = 2, buildingName = "Building B", number = "202", name = "Room 202"),
-            ),
-        )
-        val rooms = roomService.getRoomList(null, null)
-        assertEquals(2, rooms.size)
-        assertEquals("Building A", rooms[0].buildingName)
-        assertEquals("Room 101", rooms[0].name)
-        assertEquals("Building B", rooms[1].buildingName)
-        assertEquals("Room 202", rooms[1].name)
-    }
 
     @Test
     @DisplayName("강의실 목록 조회 테스트 (건물 이름 필터링)")
@@ -65,21 +46,6 @@ class RoomServiceTest {
         assertEquals("Room 101", rooms[0].name)
         assertEquals("Building A", rooms[1].buildingName)
         assertEquals("Room 102", rooms[1].name)
-    }
-
-    @Test
-    @DisplayName("강의실 목록 조회 테스트 (이름 필터링)")
-    fun getRoomListByNameTest() {
-        val name = "Room 101"
-        whenever(roomRepository.findByNameContaining(name)).thenReturn(
-            listOf(
-                Room(seq = 1, buildingName = "Building A", number = "101", name = "Room 101"),
-            ),
-        )
-        val rooms = roomService.getRoomList(null, name)
-        assertEquals(1, rooms.size)
-        assertEquals("Building A", rooms[0].buildingName)
-        assertEquals("Room 101", rooms[0].name)
     }
 
     @Test
@@ -103,9 +69,9 @@ class RoomServiceTest {
     fun getRoomByIDTest() {
         val seq = 1
         val room = Room(seq = seq, buildingName = "Building A", number = "101", name = "Room 101")
-        whenever(roomRepository.findById(seq)).thenReturn(Optional.of(room))
+        whenever(roomRepository.findByBuildingNameAndSeq("Building A", seq)).thenReturn(room)
 
-        val result = roomService.getRoomByID(seq)
+        val result = roomService.getRoomByID("Building A", seq)
         assertEquals(room, result)
     }
 
@@ -113,14 +79,14 @@ class RoomServiceTest {
     @DisplayName("강의실 ID로 조회 실패 테스트 (존재하지 않는 ID)")
     fun getRoomByIDNotFoundTest() {
         val seq = 999
-        whenever(roomRepository.findById(seq)).thenReturn(Optional.empty())
-        assertThrows<RoomNotFoundException> { roomService.getRoomByID(seq) }
+        whenever(roomRepository.findByBuildingNameAndSeq("Building A", seq)).thenReturn(null)
+        assertThrows<RoomNotFoundException> { roomService.getRoomByID("Building A", seq) }
     }
 
     @Test
     @DisplayName("강의실 생성 테스트")
     fun createRoomTest() {
-        val payload = RoomRequest(buildingName = "Building A", number = "101", name = "Room 101")
+        val payload = RoomRequest(number = "101", name = "Room 101")
         val building =
             Building(
                 id = "Y001",
@@ -128,108 +94,76 @@ class RoomServiceTest {
                 campusID = 1,
                 latitude = 37.5665,
                 longitude = 126.978,
-                url = "http://example.com/building-a",
+                url = "https://example.com/building-a",
             )
-        val room = Room(seq = null, buildingName = payload.buildingName, number = payload.number, name = payload.name)
+        val room = Room(seq = null, buildingName = "Building A", number = payload.number, name = payload.name)
 
-        whenever(buildingRepository.findByName(payload.buildingName)).thenReturn(building)
-        whenever(roomRepository.findByBuildingNameAndNumber(payload.buildingName, payload.number)).thenReturn(null)
+        whenever(buildingRepository.findByName("Building A")).thenReturn(building)
+        whenever(roomRepository.findByBuildingNameAndNumber("Building A", payload.number)).thenReturn(null)
         whenever(roomRepository.save(room)).thenReturn(room)
 
-        val result = roomService.createRoom(payload)
+        val result = roomService.createRoom("Building A", payload)
         assertEquals(room, result)
     }
 
     @Test
     @DisplayName("강의실 생성 실패 테스트 (존재하지 않는 건물 이름)")
     fun createRoomBuildingNotFoundTest() {
-        val payload = RoomRequest(buildingName = "Nonexistent Building", number = "101", name = "Room 101")
-        whenever(buildingRepository.findByName(payload.buildingName)).thenReturn(null)
+        val payload = RoomRequest(number = "101", name = "Room 101")
+        whenever(buildingRepository.findByName("Building A")).thenReturn(null)
 
-        assertThrows<BuildingNotFoundException> { roomService.createRoom(payload) }
+        assertThrows<BuildingNotFoundException> { roomService.createRoom("Building A", payload) }
     }
 
     @Test
     @DisplayName("강의실 생성 실패 테스트 (중복된 강의실 번호)")
     fun createRoomDuplicateTest() {
-        val payload = RoomRequest(buildingName = "Building A", number = "101", name = "Room 101")
-        val existingRoom = Room(seq = 1, buildingName = payload.buildingName, number = payload.number, name = "Existing Room")
+        val payload = RoomRequest(number = "101", name = "Room 101")
+        val existingRoom =
+            Room(seq = 1, buildingName = "Building A", number = payload.number, name = "Existing Room")
 
-        whenever(buildingRepository.findByName(payload.buildingName)).thenReturn(
+        whenever(buildingRepository.findByName("Building A")).thenReturn(
             Building(
                 id = "Y001",
                 name = "Building A",
                 campusID = 1,
                 latitude = 37.5665,
                 longitude = 126.978,
-                url = "http://example.com/building-a",
+                url = "https://example.com/building-a",
             ),
         )
-        whenever(roomRepository.findByBuildingNameAndNumber(payload.buildingName, payload.number)).thenReturn(existingRoom)
+        whenever(roomRepository.findByBuildingNameAndNumber("Building A", payload.number)).thenReturn(existingRoom)
 
-        assertThrows<DuplicateRoomException> { roomService.createRoom(payload) }
+        assertThrows<DuplicateRoomException> { roomService.createRoom("Building A", payload) }
     }
 
     @Test
     @DisplayName("강의실 수정 테스트")
     fun updateRoomTest() {
         val seq = 1
-        val payload = RoomRequest(buildingName = "Building A", number = "102", name = "Room 102")
+        val payload = RoomRequest(number = "102", name = "Room 102")
         val existingRoom = Room(seq = seq, buildingName = "Building A", number = "101", name = "Room 101")
-
-        whenever(buildingRepository.findByName(payload.buildingName)).thenReturn(
-            Building(
-                id = "Y001",
-                name = "Building A",
-                campusID = 1,
-                latitude = 37.5665,
-                longitude = 126.978,
-                url = "http://example.com/building-a",
-            ),
-        )
-        whenever(roomRepository.findById(seq)).thenReturn(Optional.of(existingRoom))
+        whenever(roomRepository.findByBuildingNameAndSeq("Building A", seq)).thenReturn(existingRoom)
         whenever(roomRepository.save(existingRoom)).thenReturn(
             existingRoom.apply {
-                buildingName = payload.buildingName
                 number = payload.number
                 name = payload.name
             },
         )
 
-        val result = roomService.updateRoom(seq, payload)
+        val result = roomService.updateRoom("Building A", seq, payload)
         assertEquals("Building A", result.buildingName)
         assertEquals("102", result.number)
         assertEquals("Room 102", result.name)
     }
 
     @Test
-    @DisplayName("강의실 수정 실패 테스트 (존재하지 않는 건물 이름)")
-    fun updateRoomBuildingNotFoundTest() {
-        val seq = 1
-        val payload = RoomRequest(buildingName = "Nonexistent Building", number = "102", name = "Room 102")
-        whenever(buildingRepository.findByName(payload.buildingName)).thenReturn(null)
-
-        assertThrows<BuildingNotFoundException> { roomService.updateRoom(seq, payload) }
-    }
-
-    @Test
     @DisplayName("강의실 수정 실패 테스트 (존재하지 않는 강의실 ID)")
     fun updateRoomNotFoundTest() {
         val seq = 999
-        val payload = RoomRequest(buildingName = "Building A", number = "102", name = "Room 102")
-        whenever(buildingRepository.findByName(payload.buildingName)).thenReturn(
-            Building(
-                id = "Y001",
-                name = "Building A",
-                campusID = 1,
-                latitude = 37.5665,
-                longitude = 126.978,
-                url = "http://example.com/building-a",
-            ),
-        )
-        whenever(roomRepository.findById(seq)).thenReturn(Optional.empty())
-
-        assertThrows<RoomNotFoundException> { roomService.updateRoom(seq, payload) }
+        val payload = RoomRequest(number = "102", name = "Room 102")
+        whenever(roomRepository.findByBuildingNameAndSeq("Building A", seq)).thenReturn(null)
+        assertThrows<RoomNotFoundException> { roomService.updateRoom("Building A", seq, payload) }
     }
 
     @Test
@@ -238,8 +172,8 @@ class RoomServiceTest {
         val seq = 1
         val room = Room(seq = seq, buildingName = "Building A", number = "101", name = "Room 101")
 
-        whenever(roomRepository.findById(seq)).thenReturn(Optional.of(room))
-        roomService.deleteRoom(seq)
+        whenever(roomRepository.findByBuildingNameAndSeq("Building A", seq)).thenReturn(room)
+        roomService.deleteRoom("Building A", seq)
         verify(roomRepository).delete(room)
     }
 
@@ -247,7 +181,7 @@ class RoomServiceTest {
     @DisplayName("강의실 삭제 실패 테스트 (존재하지 않는 강의실 ID)")
     fun deleteRoomNotFoundTest() {
         val seq = 999
-        whenever(roomRepository.findById(seq)).thenReturn(Optional.empty())
-        assertThrows<RoomNotFoundException> { roomService.deleteRoom(seq) }
+        whenever(roomRepository.findByBuildingNameAndSeq("Building A", seq)).thenReturn(null)
+        assertThrows<RoomNotFoundException> { roomService.deleteRoom("Building A", seq) }
     }
 }
